Initial commit
I previously made a package that simplifies 'readdir' for a
specific scenario. The use case was for software tools that need to
locate a folder that they have to access, so they need to ask to the
user its location on the first run.

I then realised that what I actually needed was just the path of
the directory and not the directory contents themselves, so this it
a package that just asks for a value (the path of the folder).

It was a bit tricky to make an implementation that I felt was
most appropriate. Once I retrieved the value, I don't want to
wait for it to be saved in a file, so a single promise wasn't
cutting it. I decided to make an event emitter with a 'then'
function resolves to the value. If I want to look for any error
before of after value retrieval, I can listen for the 'error'
event.